% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_map_data.R
\name{aggregate_map_data}
\alias{aggregate_map_data}
\title{A short description of the function}
\usage{
aggregate_map_data(portfolio)
}
\arguments{
\item{portfolio}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
