% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_portfolio.R
\name{aggregate_portfolio}
\alias{aggregate_portfolio}
\title{A short description of the function}
\usage{
aggregate_portfolio(df)
}
\arguments{
\item{df}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
