% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scenario_alignment.R
\name{calculate_scenario_alignment}
\alias{calculate_scenario_alignment}
\title{A short description of the function}
\usage{
calculate_scenario_alignment(df)
}
\arguments{
\item{df}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
