% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_technology_share.R
\name{calculate_technology_share}
\alias{calculate_technology_share}
\title{A short description of the function}
\usage{
calculate_technology_share(df)
}
\arguments{
\item{df}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
