% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weights.R
\name{calculate_weights}
\alias{calculate_weights}
\title{A short description of the function}
\usage{
calculate_weights(portfolio, portfolio_type, ...)
}
\arguments{
\item{portfolio}{A description of the argument}

\item{portfolio_type}{A description of the argument}

\item{...}{A catch-all for deprecated variables (e.g. \code{grouping_variables})}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
