% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio_subset.R
\name{create_portfolio_subset}
\alias{create_portfolio_subset}
\title{A short description of the function}
\usage{
create_portfolio_subset(portfolio, portfolio_type, ...)
}
\arguments{
\item{portfolio}{A description of the argument}

\item{portfolio_type}{A description of the argument}

\item{...}{A catch-all for deprecated variables (e.g. \code{grouping_variables})}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
