% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_utility_functions.R
\name{determine_tdm_variables}
\alias{determine_tdm_variables}
\title{A helper function to generate and keep track of all input parameters to the
\code{calculate_tdm} function.}
\usage{
determine_tdm_variables(start_year)
}
\arguments{
\item{start_year}{An integer value, indicating the start year of the PACTA
run.}
}
\value{
A named list containing all relevant inputs to \code{calculate_tdm()}
}
\description{
A helper function to generate and keep track of all input parameters to the
\code{calculate_tdm} function.
}
