% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abcd_raw.R
\name{get_abcd_raw}
\alias{get_abcd_raw}
\title{A short description of the function}
\usage{
get_abcd_raw(portfolio_type, data_path, start_year, time_horizon, sector_list)
}
\arguments{
\item{portfolio_type}{A description of the argument}

\item{data_path}{Path to directory containing ABCD Data files}

\item{start_year}{Start year of the analysis}

\item{time_horizon}{Length of the analysis in years (analysis will run from
\code{start_year} to \code{start_year + time_horizon})}

\item{sector_list}{(character) A vector of PACTA sectors to include in the
analysis.}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
