% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_abcd_from_db.R
\name{merge_abcd_from_db}
\alias{merge_abcd_from_db}
\title{Merge ABCD_scenario data from a database file (RDS or SQLite) with the
portfolio data}
\usage{
merge_abcd_from_db(
  portfolio,
  portfolio_type,
  db_dir,
  equity_market_list,
  scenario_sources_list,
  scenario_geographies_list,
  sector_list,
  id_col = "id"
)
}
\arguments{
\item{portfolio}{A data frame containing the processed portfolio data}

\item{portfolio_type}{A single string specifying the type of assets in the
portfolio}

\item{db_dir}{A single string specifying the path to the directory that
contains the database file}

\item{equity_market_list}{A character vector of equity markets to be included}

\item{scenario_sources_list}{A character vector of scenario sources to be
included}

\item{scenario_geographies_list}{A character vector of senario geographies to
be included}

\item{sector_list}{A character vector of sectors to be included}

\item{id_col}{A single string specifying the name of the id column in the
portfolio data}
}
\value{
A data frame contianing the portfolio data with the merged
ABCD_scenario data
}
\description{
Merge ABCD_scenario data from a database file (RDS or SQLite) with the
portfolio data
}
