% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_in_geography.R
\name{merge_in_geography}
\alias{merge_in_geography}
\title{A short description of the function}
\usage{
merge_in_geography(portfolio, ald_raw)
}
\arguments{
\item{portfolio}{A description of the argument}

\item{ald_raw}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
