% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownership_allocation.R
\name{ownership_allocation}
\alias{ownership_allocation}
\title{A short description of the function}
\usage{
ownership_allocation(portfolio)
}
\arguments{
\item{portfolio}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
