% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_weight_allocation.R
\name{port_weight_allocation}
\alias{port_weight_allocation}
\title{A short description of the function}
\usage{
port_weight_allocation(port_abcd)
}
\arguments{
\item{port_abcd}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
