% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_calculations.R
\name{pw_calculations}
\alias{pw_calculations}
\title{A short description of the function}
\usage{
pw_calculations(eq_portfolio, cb_portfolio, ...)
}
\arguments{
\item{eq_portfolio}{A description of the argument}

\item{cb_portfolio}{A description of the argument}

\item{...}{A catch-all for deprecated variables (e.g. \code{grouping_variables})}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
