% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_conditions_met.R
\name{tdm_conditions_met}
\alias{tdm_conditions_met}
\title{A helper function to determine what conditions should be met for the TDM to
be generated}
\usage{
tdm_conditions_met(
  data,
  t0,
  delta_t1,
  delta_t2,
  scenarios,
  project_code,
  additional_groups = NULL
)
}
\arguments{
\item{data}{Like the \code{data} input to \code{calculate_tdm}.}

\item{t0}{Like the \code{t0} input to \code{calculate_tdm}.}

\item{delta_t1}{Like the \code{delta_t1} input to \code{calculate_tdm}.}

\item{delta_t2}{Like the \code{delta_t2} input to \code{calculate_tdm}.}

\item{scenarios}{Like the \code{scenarios} input to \code{calculate_tdm}.}

\item{project_code}{The PACTA project code.}

\item{additional_groups}{Like the \code{additional_groups} input to \code{calculate_tdm}.}
}
\value{
A Boolean, indicating if the TDM should be calculated.
}
\description{
A helper function to determine what conditions should be met for the TDM to
be generated
}
