% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_utility_functions.R
\name{tdm_scenarios}
\alias{tdm_scenarios}
\title{A helper function to keep track of the scenario(s) relevant for the TDM.}
\usage{
tdm_scenarios()
}
\value{
A character vector, containing the \code{scenario} value(s) relevant to the TDM.
}
\description{
A helper function to keep track of the scenario(s) relevant for the TDM.
}
