% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abcd_raw.R
\name{get_abcd_raw}
\alias{get_abcd_raw}
\title{A short description of the function}
\usage{
get_abcd_raw(portfolio_type, supervisor_workflow = FALSE)
}
\arguments{
\item{portfolio_type}{A description of the argument}

\item{supervisor_workflow}{A description of the argument}
}
\value{
A description of the return value
}
\description{
A longer description of the function
}
